requirements.txt
psutil>=5.9.0
Pillow>=10.0.0
fonttools>=4.0.0
